Refactor large plotHelpers file
