+++ api/utils/queryBuilder.mjs
@@ -73,7 +73,7 @@ export const getGroupedCountsByIdentifiers = ({ ids, idRanges, idColumn, groupBy
     let remappedGroupBy = handleIdKeyIrregularities(groupBy, table);
     clauses = `${clauses.length > 0 ? `WHERE ${clauses.join(' OR ')}` : ''}`;
     return `
-        SELECT ${remappedGroupBy} as name, COUNT(*) as count, (COUNT(*) * 100.0 / SUM(COUNT(*)) OVER()) as percent, SUM(mbases)/POWER(10,3) as gbp
+        SELECT ${remappedGroupBy} as name, COUNT(*) as count, SUM(mbases)/POWER(10,3) as gbp
         FROM ${table}
         ${clauses}
         GROUP BY ${remappedGroupBy}
