+++ frontend/src/features/Figures/SRARun/SRARunLayout.tsx
